// src/app/pages/customers/models/item-price-customer.ts
export class ItemPriceCustomerModel {
    itemId: number;
    itemCode: string;
    itemName: string;
    priceList1: number;
    priceList2: number;
    priceList3: number;
    priceList4: number;
    priceList5: number;
    priceList6: number;
    priceList7: number;
    priceList8: number;
    priceList9: number;
    priceList10: number;
    priceList1Enabled: boolean;
    priceList2Enabled: boolean;
    priceList3Enabled: boolean;
    priceList4Enabled: boolean;
    priceList5Enabled: boolean;
    priceList6Enabled: boolean;
    priceList7Enabled: boolean;
    priceList8Enabled: boolean;
    priceList9Enabled: boolean;
    priceList10Enabled: boolean;
    isModified: boolean;
    listPriceId: number;
    price: number;
}